+++ react/AffiliateProfileTotalizer.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import AffiliateProvider from './context/AffiliateProvider'
+import ProfileTotalizer from './components/store/ProfileTotalizers'
+
+function AffiliateProfileTotalizer() {
+  return (
+    <AffiliateProvider>
+      <div>
+        <ProfileTotalizer />
+      </div>
+    </AffiliateProvider>
+  )
+}
+
+export default AffiliateProfileTotalizer